Move NEON-specific SIMD helpers from `simd.hpp` to `neon.hpp`
